# rest-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/search/autocomplete" : {
      "get" : {
        "tags" : [ "Search" ],
        "operationId" : "AutoComplete",
        "parameters" : [ {
          "name" : "term",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  },
                  "x-content-type" : "application/json"
                },
                "examples" : {
                  "default" : {
                    "value" : "[\"term1\", \"term3\", \"term6\", \"term4\"]"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tags" : {
      "get" : {
        "tags" : [ "Tags" ],
        "operationId" : "GetTags",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "full_perms",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_16"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":2,\"next\":null,\"previous\":null,\"all\":[2,1],\"results\":[{\"id\":2,\"slug\":\"deltag\",\"name\":\"deltag\",\"color\":\"#b62d74\",\"text_color\":\"#ffffff\",\"match\":\"test\",\"matching_algorithm\":5,\"is_insensitive\":true,\"is_inbox_tag\":false,\"document_count\":0,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}},{\"id\":1,\"slug\":\"testtag\",\"name\":\"testtag\",\"color\":\"#ddd977\",\"text_color\":\"#000000\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"is_inbox_tag\":true,\"document_count\":1,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}]}"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Tags" ],
        "operationId" : "CreateTag",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/api_tags_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"deltag\",\"color\":\"#b62d74\",\"is_inbox_tag\":false,\"matching_algorithm\":5,\"match\":\"test\",\"is_insensitive\":true,\"owner\":3}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_17"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"deltag\",\"name\":\"deltag\",\"color\":\"#b62d74\",\"text_color\":\"#ffffff\",\"match\":\"test\",\"matching_algorithm\":5,\"is_insensitive\":true,\"is_inbox_tag\":false,\"owner\":3,\"user_can_change\":true}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tags/{id}" : {
      "put" : {
        "tags" : [ "Tags" ],
        "operationId" : "UpdateTag",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/tags_id_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"id\":2,\"slug\":\"deltag\",\"name\":\"deltag2\",\"color\":\"#b62d74\",\"text_color\":\"#ffffff\",\"match\":\"test\",\"matching_algorithm\":5,\"is_insensitive\":true,\"is_inbox_tag\":false,\"document_count\":0,\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"permissions_form\":{\"owner\":3,\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_18"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"deltag2\",\"name\":\"deltag2\",\"color\":\"#b62d74\",\"text_color\":\"#ffffff\",\"match\":\"test\",\"matching_algorithm\":5,\"is_insensitive\":true,\"is_inbox_tag\":false,\"document_count\":0,\"owner\":3,\"user_can_change\":true}"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Tags" ],
        "operationId" : "DeleteTag",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "apicorrespondentsid_permissions_view" : {
        "required" : [ "groups", "users" ],
        "type" : "object",
        "properties" : {
          "users" : {
            "type" : "array",
            "items" : { }
          },
          "groups" : {
            "type" : "array",
            "items" : { }
          }
        },
        "example" : {
          "groups" : [ "", "" ],
          "users" : [ "", "" ]
        }
      },
      "apicorrespondentsid_permissions" : {
        "required" : [ "change", "view" ],
        "type" : "object",
        "properties" : {
          "view" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions_view"
          },
          "change" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions_view"
          }
        },
        "example" : {
          "view" : {
            "groups" : [ "", "" ],
            "users" : [ "", "" ]
          },
          "change" : null
        }
      },
      "apicorrespondentsid_permissions_form" : {
        "required" : [ "owner", "set_permissions" ],
        "type" : "object",
        "properties" : {
          "owner" : {
            "type" : "integer"
          },
          "set_permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          }
        }
      },
...
      "inline_response_200_16" : {
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "type" : "array",
            "items" : {
              "type" : "integer"
            }
          },
          "results" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_16_results"
            }
          }
        },
        "example" : {
          "next" : 6,
          "all" : [ 5, 5 ],
          "previous" : 1,
          "count" : 0,
          "results" : [ {
            "owner" : 9,
            "matching_algorithm" : 2,
            "document_count" : 7,
            "color" : "color",
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              },
              "change" : null
            },
            "name" : "name",
            "match" : "match",
            "id" : 5,
            "text_color" : "text_color",
            "is_inbox_tag" : true,
            "slug" : "slug"
          }, {
            "owner" : 9,
            "matching_algorithm" : 2,
            "document_count" : 7,
            "color" : "color",
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ "", "" ]
              },
              "change" : null
            },
            "name" : "name",
            "match" : "match",
            "id" : 5,
            "text_color" : "text_color",
            "is_inbox_tag" : true,
            "slug" : "slug"
          } ]
        }
      },
      "api_tags_body" : {
        "required" : [ "color", "is_inbox_tag", "is_insensitive", "match", "matching_algorithm", "name", "owner" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "match" : {
            "type" : "string"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "integer"
          }
        }
      },
      "inline_response_200_17" : {
        "required" : [ "color", "id", "is_inbox_tag", "is_insensitive", "match", "matching_algorithm", "name", "owner", "slug", "text_color", "user_can_change" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "text_color" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "integer"
          },
          "user_can_change" : {
            "type" : "boolean"
          }
        },
        "example" : {
          "owner" : 1,
          "matching_algorithm" : 6,
          "user_can_change" : true,
          "color" : "color",
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "text_color" : "text_color",
          "is_inbox_tag" : true,
          "slug" : "slug"
        }
      },
      "tags_id_body" : {
        "required" : [ "color", "document_count", "id", "is_inbox_tag", "is_insensitive", "match", "matching_algorithm", "name", "owner", "permissions", "permissions_form", "set_permissions", "slug", "text_color" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "text_color" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          },
          "set_permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          },
          "permissions_form" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions_form"
          }
        }
      },
      "inline_response_200_18" : {
        "required" : [ "color", "document_count", "id", "is_inbox_tag", "is_insensitive", "match", "matching_algorithm", "name", "owner", "slug", "text_color", "user_can_change" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "text_color" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "user_can_change" : {
            "type" : "boolean"
          }
        },
        "example" : {
          "owner" : 5,
          "matching_algorithm" : 6,
          "user_can_change" : true,
          "document_count" : 1,
          "color" : "color",
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "text_color" : "text_color",
          "is_inbox_tag" : true,
          "slug" : "slug"
        }
      },
...
      "inline_response_200_16_results" : {
        "required" : [ "color", "document_count", "id", "is_inbox_tag", "is_insensitive", "match", "matching_algorithm", "name", "owner", "permissions", "slug", "text_color" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "text_color" : {
            "type" : "string"
          },
          "match" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer"
          },
          "owner" : {
            "type" : "integer"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/apicorrespondentsid_permissions"
          }
        },
        "example" : {
          "owner" : 9,
          "matching_algorithm" : 2,
          "document_count" : 7,
          "color" : "color",
          "is_insensitive" : true,
          "permissions" : {
            "view" : {
              "groups" : [ "", "" ],
              "users" : [ "", "" ]
            },
            "change" : null
          },
          "name" : "name",
          "match" : "match",
          "id" : 5,
          "text_color" : "text_color",
          "is_inbox_tag" : true,
          "slug" : "slug"
        }
      },
...
    }
  }
}